+++ src/raygun.breadcrumbs.js
@@ -6,7 +6,7 @@
  * raygun4js
  * https://github.com/MindscapeHQ/raygun4js
  *
- * Copyright (c) 2017 MindscapeHQ
+ * Copyright (c) 2024 MindscapeHQ
  * Licensed under the MIT license.
  */
 /* globals console */
@@ -475,7 +475,7 @@ window.raygunBreadcrumbsFactory = function(window, Raygun) {
       error.duration = error.duration + 'ms';
       self.recordBreadcrumb({
         type: 'request',
-        message: 'Failed request to ' + error.requestUrl,
+        message: 'Failed request to ' + error.requestURL,
         level: 'info',
         metadata: error,
       });